% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_session.R
\name{by_session}
\alias{by_session}
\title{Calculates session duration and other measures}
\usage{
by_session(frame)
}
\arguments{
\item{frame}{Data frame in standard format for this system

Each session is about a particular document}
}
\description{
Calculates session duration and other measures
}
