% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Translate a raw submission file into the standard format}
\usage{
get_submissions(csv_name = NULL, key = NULL)
}
\arguments{
\item{csv_name}{Name of the CSV file.}

\item{key}{The google key of the sheet holding submissions}
}
\description{
Translate a raw submission file into the standard format
}
\details{
Provide either of \code{csv_name} or \code{key}, but not both.
}
