% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions_google_forms.R
\name{get_submissions_google_forms}
\alias{get_submissions_google_forms}
\title{Reading Hash submissions from a GoogleForms sheet}
\usage{
get_submissions_google_forms(csv_name)
}
\description{
Reading Hash submissions from a GoogleForms sheet
}
