% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_an_item.R
\name{summarize_MC_item}
\alias{summarize_MC_item}
\title{Tabulate answers for one question item}
\usage{
summarize_MC_item(events, item_name, fiducial_date = 0.25)
}
\arguments{
\item{events}{Events in the standard format}

\item{item_name}{Name of the item}

\item{fiducial_level}{A quantile level for the fiducial date. Default: 0.25}
}
\description{
Tabulate answers for one question item
}
\details{
\itemize{
\item \code{first_response} gives the time (in hours) from the student's first
answer to the question.
\verb{correct_response`` gives the time from the students first correct response }last_response`, similar but for the last response, true or not
}
}
